--- src/components/atoms/register/BusyIndicator/Styles.js
@@ -1,24 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const LoadingIndicatorStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.4)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  background: {
-    width: 120,
-    height: 120,
-    backgroundColor: '#20232A',
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: 15,
-  },
-  message: {
-    marginTop: 5,
-    color: '#33A8FF',
-  },
-});
-
-export default LoadingIndicatorStyles;
