--- src/components/atoms/register/checkbox/Style.js
@@ -1,26 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const Style = StyleSheet.create({
-  checkboxContainer: {
-    flexDirection: 'row',
-  },
-  label: {
-    marginTop: 6,
-  },
-  error: {
-    color: 'red',
-    fontSize: 13,
-    paddingBottom: 1,
-    marginHorizontal: 6,
-  },
-  iconerror: {
-    marginTop: 2,
-    color: 'red',
-    marginLeft: 10,
-  },
-  Containerow: {
-    flexDirection: 'row',
-  },
-});
-
-export default Style;
